Day 0: Initial planning and thoughts
Laid out my initial planning and thoughts for the this challenge. I want to build some strong frontend skills and diversify my toolset with a couple different technologies.

My goal with this challenge is to become a solid frontend developer by understanding the fundamentals of modern frontends and how to build complex and efficient applications.